// types/datepicker.ts
export enum RecurrenceTypes {
  DAILY = "DAILY",
  WEEKLY = "WEEKLY",
  MONTHLY = "MONTHLY",
  YEARLY = "YEARLY",
  CUSTOM = "CUSTOM",
}

export enum RepeatTypes {
  CUSTOM = "CUSTOM",
  ENDLESS = "ENDLESS",
}

export enum FrequencyTypes {
  DAY = "DAY",
  WEEK = "WEEK",
  MONTH = "MONTH",
  YEAR = "YEAR",
}